backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "my-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "psycopg2-binary",
    "python-dotenv",
    "websockets",
    "requests",
    "apscheduler",
    "pandas>=2.3.3",
    "ccxt>=4.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["main.py"]

